New version! Move to .sublime-completion file
It was hard to get an overview of the snippets collection; what was
included, if there were any duplicate triggers etc. To remedy this,
everything is now moved to a sublime-completions file.

Some missing properties and additional shortcuts have been added, and
duplicate triggers are fixed. The !important snippet file is there to
allow the exclamation mark (!) as a trigger (though I never tried
escaping it in the completions file, oh well).